stage0Module3TaskeFinal

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,8 +3,11 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius) {
 
-        float fahrenheit = (float) (temperatureCelsius * 9 / 5) + 32f;
+         float fahrenheit = ((temperatureCelsius * 9) / 5) + 32;
+
+
         System.out.println(fahrenheit);
+
 
 
     }
